Ast: Fix usage of dangling pointer to root node
If the nodes list was resized, then the pointer to the root could no longer be valid, and the stored pointer was being compared when finalizing the tree. This meant that if the root node was no longer at that address, it would attempt to call `isClosed()` on the root which hits unreachable.