fmt: automatically remove self-closing slash from void elements
Non-void elements will not be automatically fixed and the error will
remain for the user to see. One reason for this behavior is to make
sure we don't "sneakily" remove a self-closing tag from an element that
was hiding a nesting error. By having the user face the problem they
get a chance to discover the latent mistake and fix it, on top of
learning a bit more about html.

This is what I mean by latent nesting bug:

```
<body>
  <div />
  <div>
</body>
```
The indentation of this code seems to suggest that the two divs are
sibling, while in reality the second one is a child of the first because
in HTML self-closing slashes are ignored.

As of today editors like VSCode will produce misleading indentation
in the presence of a self-closing slash, helping perpetuate mistakes.